Extract jpackage conventions
Reduces the size of qupath-app (more refactoring is still needed).
